Including more recent Python versions
Including more recent Python versions in the classifiers, as used in PyPI releases. Also added a generic Python classifier.

While I've not used the current cbor 1.0.0 with Python 3.6, it does work (for me, not fully tested...) with Python 3.7 (v3.7.3 in fact) when installing it via PIP on a Raspberry Pi with Rasbian Buster; therefore, I assumed the 3.6 and 3.7 releases should work fine as well.

As said, not fully tested; so it is perfectly fine to ignore this PR for that reason. No hard feelings, no need to explain.