Fix issue #6 broken test_sortkeys()
Newer python dictionaries are keeping the order the keys were added in.
As a result since both lists were pre-sorted and identicle so the later
test failed to find any expected differences causing the failure.
This change introduces 3 entries that are out of sorted order for both
the obytes and xbytes lists.  This ensures a valid test of sorting.